fix queries
implementation scooters with more than X travels in a year,fix bugs of queries in reports.

+++ travel-microservice/src/main/java/org/example/travelmicroservice/dtos/TravelReportDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public class TravelReportDTO {
 
     private Long travelId;
-    private Integer scooterId;
+    private Long scooterId;
     private LocalDate date;
     private String effectiveUsageTime;
 }
